requirements.txt
"""
# Core dependencies
asyncio==3.4.3
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2

# Database & Cache
asyncpg==0.29.0
redis==5.0.1
sqlalchemy[asyncio]==2.0.23

# Exchanges
ccxt>=4.4.0
requests==2.31.0
websockets==12.0

# Monitoring & Alerts
prometheus-client==0.19.0
python-telegram-bot==20.7

# Math & Analysis
numpy==1.26.2
pandas==2.1.4
"""